Work In Progress
Added code to integrate the OpenSSL Library. Wrote code to create an HTTPS connection through SSL and catch errors in the connection.
HTTP connections using C standard libraries is included in comments, but not fit for ant HTTPS connections.
Current connection is to xkcd.com/rss.xml but the connection is failing. More troubleshooting is needed.